Rename js files
